Adding the -std=c99 flag to our generated code for the GNU tool suite which cleanly takes care of the log2() and exp2() problems we were seeing before (and should also keep us a bit more honest in our generated code, as it turns out -- un-prototyped routines result in errors with this flag, which is nice).
On Sun, for some reason, I had to undefine a macro
in order to get routines like strtoull made available
to the compiler -- not sure why.  Either our Suns
are poorly installed (which I'm guessing), or there's
some gotcha on Suns which I have yet to run into.
We'll see what tonight's regressions do.

Added additional prototypes to the runtime to get
around the warnings generated by these new flags
and removed the bogus declarations of log2() and
exp2() in chplmath.h.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10688 3a8e244f-b0f2-452b-bcba-4c88e055c3ca