Fix trivial/marybeth/typeselect.chpl future for return statements in type select statements.
Previously, the implementation was to generate one function per when statement
with arguments whose types match the types for that case that executes the
statements for that case.

Now, the functions are parameter functions that return an integer, and a
normal select statement over the integer values is used.  Since the functions
are params, all of the incorrect cases are folded leaving only the one that
matches, with zero runtime overhead.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10798 3a8e244f-b0f2-452b-bcba-4c88e055c3ca