Added code to mangle formal argument names if the names duplicate one another.
Added a temp to replace the implicit "setter" argument with true or
false when references are culled in case the address of this argument
is taken.  This happens in the case that it is passed to a nested
function.  Previously, the replacement was direct.

This fixes release/examples/sparse with --no-inline.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10849 3a8e244f-b0f2-452b-bcba-4c88e055c3ca