Added two more regressions for getpwuid calls in the compiler for valgrind testing related to compflags/bradc/gdbddash/declint
Added memory count files for chap09 and chap10.

This should leave only one remaining linux64 valgrind regression
for the test that purposefully runs out of memory (which causes
valgrind to print out a bunch of different stuff).  I need to
investigate whether there's a graceful way to handle this, but
probably not today.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10883 3a8e244f-b0f2-452b-bcba-4c88e055c3ca