Removed --parallel,-p flag as it is now redundant with --serial and not as graceful.
Moved work in parallel1 into cleanup (and simplified) as what is done
here is exactly what cleanup is intended for--minor straightening that
is more difficult to do while parsing.  Indeed, we may even want to
move this work into parsing as it would be simple to do there too.
Cleaned the code in parallel2 (now parallel) pass.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10887 3a8e244f-b0f2-452b-bcba-4c88e055c3ca