This is a stopgap measure to allow one to call a C function from Chapel.  It is crufty at best.  I've been sitting on this work for a while debating on whether to checkin but I accidentally integrated it with other work so am trying to tease it out.
Added support for calling C functions from Chapel by prototyping the C
function as if it were a Chapel function.  No overloading is allowed.
Support is only provided for primitive types.

Use the new keyword _extern before the function prototype.

To compile the C code, or link the C object file, add it to an option
using the --ccflags.

As an example, see the new test link/deitz/test_link1.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10896 3a8e244f-b0f2-452b-bcba-4c88e055c3ca