Refactored the runtime to make its Makefile structure mirror that of the compiler.
Added sibling directories to src/ called threads-none and
threads-pthreads in the start of an effort to factor pthreads out of
the threading support that we provide.  This makes pthreads one of
many alternatives, though the only one currently.

Note that src/ should be renamed common/ or serial/ as this is the
serial runtime support files.

Added another environment variable CHPL_THREADS that can be used to
specify a different threading package.  Currently, the default is
pthreads though we may instead want to require this be set.  Also,
pthreads is currently the only supported threading package.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10905 3a8e244f-b0f2-452b-bcba-4c88e055c3ca