Continually haunted by the fact that:    chpl --help --no-devel and:    chpl --no-devel --help return different results due to the fact that help is implemented in our flag mechanism as a function callback rather than a change of state, I rejiggered the printing of: --help, --copyright, --license, --version to avoid the use of these functions and print things out a bit more cleanly and uniformly.
Also, changed output stream for this information
from stderr to stdout (to match what other compilers
do), and made the status message for --help depend
on whether the user requested it explicitly or it
was the result of not supplying a file argument.

Inserted some new tests to lock in behavior if
pairs of flags or all four flags are used.  Removed
old --version test which had become redundant w.r.t.
this stuff; left userhelp.chpl as it was to avoid
confusion (it could just have easily been moved
into this new printstuff directory).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10929 3a8e244f-b0f2-452b-bcba-4c88e055c3ca