Future:  The compiler generates an internal error during function resolution on PRIMITIVE_SET_REF when it sees a record that contains an array of records with the same type (record).
It seems like this would have to be an error when the array is defined with a
default size, since this would result in an infinite size allocation.  However,
for arrays that are defined to be [1..0] this seems like it should be okay
assuming that arrays are stored by reference in records.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10930 3a8e244f-b0f2-452b-bcba-4c88e055c3ca