Future: using a module in your program forces you to define main().
When you write "use module;" in your program, it is not possible to have the
program's entry point be outside of any function.  This results in:

error: Code defines multiple modules but no main function.

The future example is condensed to a single source file, but I also get the
same error if I put the "use module; code...;" into a separate file from the
module.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10956 3a8e244f-b0f2-452b-bcba-4c88e055c3ca