This is the random walk sudoku solver I've been babbling about.  The code is a little frightening since it was written fast and hacked up to use less memory and go fast.
* sudoku.chpl is only the random walk simulated annealing solver.

* sudoku-smart.chpl attempts to infer as many spots as possible before invoking
the above solver.  In easy to hard puzzles, the solution is found without
performing random search, however the evil and outlaw puzzles require more
sophisticated inference and in some cases guessing, so they invoke the random
solver.

Warning: The evil and outlaw puzzles require a lot of memory to solve.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10972 3a8e244f-b0f2-452b-bcba-4c88e055c3ca