Added initial support for reading and writing local variables and arguments within on statements.
Added a new pass called insertFatPointers that inserts fat pointers
"_chpl_comm_ptr_t" to remote data and passes a wrapper class of these
fat pointers to the function that is forked for handling on
statements.

I would be happy to go over this checkin with anybody interested as it
starts to lay some framework for much future work.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11006 3a8e244f-b0f2-452b-bcba-4c88e055c3ca