In this check-in, we throw a few more warning flags (motivated by those used by the GASNet installation) to keep our compiler, runtime, and generated code a bit cleaner.  The upshot of this is that we now compile gasnet code warning free (they were throwing additional flags on our generated code which could cause panic among the faint-hearted like myself).
The flags added were:

For C++: -Wno-unused -Wpointer-arith -Wwrite-strings
For C:   -Wnested-externs -Wdeclaration-after-statement

The only one of these that had any significant effect
on our code (and it was significant) was -Wwrite-strings.
This flag causes warnings for things like:

  char* str = "hello!";

and:

  void foo(char* ) { ... }
  foo("hello!");

because in both cases, the string "hello!" is a
const char* (this, I believe is a fairly new addition
to the C spec) meaning that the characters can't be
overwritten (and in particular may be implemented by
being put into static memory).  So by assigning them
or passing them to a plain old char*, we're potentially
doing something unsafe since if that variable's
character buffer is written to later, we'd be modifying
this constant thing.

As it turns out, we almost always use const char*'s
in our code because we rarely overwrite a buffer of
characters, but rather create a new one.

There was a fair amount of work required to convert all
our uses of char*'s (which cascaded throughout our
functions and classes in a major way), but having
converted it, it should (a) be easy to maintain, and
(b) help us write cleaner, more precise, less
error-prone code.

That's the idea anyway.  If it gives people a headache,
please give a shout.  Ditto if there are any questions
about the kinds of warnings this flag results in and
how to fix them.   I'm now quite experienced.

I think the only case where I found a string that was
actually being modified was a case where we made a
string that was logically splicing one string into
another and made the insertion point into '\0' as a
little trick for making this work.  I replaced it
with another trick that didn't rely on modifying the
string.

The other main change that one might see here is that
the type of _string in the generated code is now
const char* rather than char*.  This seems appropriate
given how we use them, at least in the current
implementation (and it seems to work).  We can reconsider
later if necessary.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11043 3a8e244f-b0f2-452b-bcba-4c88e055c3ca