Fixed regressions in David's config param flag tests thanks to some incredibly helpful debugging from David (it would've taken me forever to find this).
Because Immediate() supports char* and bool constructors,
and because we were passing a const char* in, C++ decides
the better match is the bool constructor.  Oog.  Fixed
this by making the Immediate constructor that takes a
char* take a const char* and propagating the const-ness
to related variables, methods, and functions similar to
what the bulk of my check-in yesterday did.

Thanks David!


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11048 3a8e244f-b0f2-452b-bcba-4c88e055c3ca