Added ability to specify a commlayer-specific .good file to the testing system, using the name:
   foo.comm-$CHPL_COMM.good

Typically, comm-none should be used to specify such files
assuming that the default output would be similar for any
comm layer that supports multiple locales.  As a result
of this change, the current order of precedence for .good
files is:

      foo.<machine name>.good
      foo.comm-$CHPL_COMM.good
      foo.<platform>.good
      foo.good

Obviously, one might want a richer hierarchy combining
machine name and/or platform with comm layer, but the
current hierarchy seemed like it would take us quite far,
so I didn't implement anything more complicated for now.

Updated comments in the start_test file to talk about
this and to provide a bit more detail about using numlocales
files.

Updated my numLocales tests so that they should now work
correctly with either gasnet or none as CHPL_COMM.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11158 3a8e244f-b0f2-452b-bcba-4c88e055c3ca