Added macros isDerivedClass(ast) for all derived classes of BaseAST. These macros return true if ast points to a dynamic subclass of BaseAST.
Added macros toDerivedClass(ast) to safely downcast the ast pointer.
This is equivalent to dynamic_cast<DerivedClass*>(ast), but is
significantly faster.

Optimized the cleanAst routine, called between passes, to avoid use of
dynamic_cast and to no longer clear use and def vectors.  This drops
the time between passes on Jacobi from 1.9 to 1.2 seconds with the
vast majority of time between passes spent in the verify code rather
than the clean code.

Added an INT_FATAL to check for the case where addPragma or addPragmas
is called but no pragma is actually added because 'this' is not a
DefExpr or a Symbol.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11234 3a8e244f-b0f2-452b-bcba-4c88e055c3ca