Checking in a future regarding variable length argument lists using homogeneous tuples.
In this test, I attempted to query the type of the elements of
a variable length argument list.  Instead, I got a tuple type.

I think this is a bug.  If the tuple type is the correct type
for this type query, then how do I access the type of the tuple
elements?


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11235 3a8e244f-b0f2-452b-bcba-4c88e055c3ca