Made for_alist macro take only two arguments and declare the first argument to be of type Expr.  So
  for_alist(expr, exprlist) { ...

is equivalent to the old for_alist written as

  for_alist(Expr, expr, exprlist) { ...

but is implemented without an unnecessary dynamic_cast.

Ditto with for_alist_backward.

Added new macros for_defs and for_defs_backward that is equivalent to
the old for_alists written as

  for_alist(DefExpr, def, defexprlist) { ...
    for_alist_backward(DefExpr, def, defexprlist) { ...

Removed dynamic_casts from all of the stylized loops over alists,
replacing them with the toBaseAST macros.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11237 3a8e244f-b0f2-452b-bcba-4c88e055c3ca