Added a flush() call to the Writer class and made our write() call invoke flush() after writing out all of its arguments but before releasing the file's lock.  This seems to fix the races on output not getting flushed before GASNet exits, though I'm not sure why this would have any effect.  Am continuing to follow up with the GASNet team to try and understand this.
This may have a long-term performance impact if we end up
wanting/needing to minimize flush() calls.  In such an
event, we should probably expose a flush() call to the user
and clarify the semantics of when a flush() call will be
invoked on the user's behalf.  For now, I think flushing
on every write is reasonable.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11261 3a8e244f-b0f2-452b-bcba-4c88e055c3ca