Fixed a bug in which trying to give the Chapel compiler a directory didn't result in any sort of warning.  For some reason, fopen() doesn't seem to complain about this even though at least one web reference says it should. Maybe the standard changed recently.
So now, I try to read a character from each input file
as part of our "make sure these are all files" check
which already existed.  If I can't read a character,
I report an error complaining that the file is either
empty or that it's a directory.  If anyone knows of a
portable way to distinguish these two, let me know.

Added emptyFile.chpl to lock in this behavior, even
though I'd be happy to have it accept the empty file
or do something less "either a or b is wrong".


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11327 3a8e244f-b0f2-452b-bcba-4c88e055c3ca