Implemented compress() and reconstruct() that converts the numerical representation between the scaling and multiwavelet bases using the two-scale relationship.  Also implemented summarize() that outputs the 2-norms of the sum and difference coefficients in each level of the function tree for debugging.
I ran into a bug in associative domains where things go bad removing something
from the domain and then trying to add it back in.  Worked around it for now by
adding valid flags to everything in the array so nothing has to be deleted.  I
plan to tackle this further tomorrow.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11365 3a8e244f-b0f2-452b-bcba-4c88e055c3ca