Fix 3 --baseline regressions.  Mark getter functions as compiler temps and when functions get cloned, make sure the isCompilerTemp field gets copied.
Also, fix a future about modules that would have name collisions if in the same
scope, but at different scopes. At this point, the only problem was the bool
value that prevents module initialization functions from being run multiple
times. Mangle the names of those symbols by appending a number to the end.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11386 3a8e244f-b0f2-452b-bcba-4c88e055c3ca