Fixing a few more associative domain/array bugs: (1) accessing an associative array out-of-bounds now results in an out-of-bounds halt message; (2) removing an index from an associative array results in a similar halt message.
Fixed up some tests of Wayne's that assumed that
an index could be removed from an indefinite
domain even if it wasn't in there (I can't think
of a reason to support this presently, and the
spec doesn't say either way, so would prefer to
throw the error until that creates a problem).

Made slight mods to Jim's motivating tests to
have them print the domain/array if a halt
isn't reached, just as a sanity check that
they're working.

Removed Aniruddha associative array future which
my previous check-in fixed.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11411 3a8e244f-b0f2-452b-bcba-4c88e055c3ca