[One expected regression related to the test script trying to compile a CVS directory.  I have a fix, but am currently testing it).
Adding support for specifying .c and .o files on the
Chapel command line, rather than using --ldflags as
a means of passing a .c or .o file to the linker.
(This caused problems for builds, like GASNet, where
the compiler and linker commands are different; in
particular, for GASNet, gcc is done for the compile
steps, g++ for the link steps, so specifying a .c file
on the link line, munges its names in g++ identifiers,
making the link step fail (for Steve's test_link1.chpl
test, for example).

In making this change, also required Chapel source
files to have the .chpl extension, and throw an error
otherwise, since we are now basing actions on the
input files on the types of the files.

Updated tests that used --ldflags with .c files to
simply use the .c files.  Added a test of compiling
with an unknown suffix.  Updated my "readDirInsteadOfFile"
test to use a directory with a .chpl name instead of
an extension-less name (since that now generates an
"unknown extension" error).

Didn't add a test of a .o file on the command line
because I couldn't figure out a way to generate a
.o file in a sufficiently generic way for all
platforms (without lots of inclusion of the normal
Makefiles, which scared me).

Note that the way I implemented this was to generate
rules for converting the user's .c files into .o
files in order to avoid dumping .o files into the
user's directories, which was the most straightforward
way to do it in Makefiles.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11502 3a8e244f-b0f2-452b-bcba-4c88e055c3ca