Modified the conditions used to infer the return type of a function with multiple return statements so that instead of only looking at coercions, it also looks at the class hierarchy.
Given classes

  class C {}
  class D1: C {}
  class D2: C {}

this change infers a return type of C if a function returns a C on one
path but a D1 on another.  Removed a future.

On the other hand, if a function returns a D1 on one path and a D2 on
another, the result is an error: the compiler cannot infer the return
type.  Changed a good and removed a future.

For this other hand, our plan to return the most immediate base class
failed since we are not a single-inheritance language.

Updated the spec to discuss return types; previously it did not.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11514 3a8e244f-b0f2-452b-bcba-4c88e055c3ca